fdt_parser::base

Struct InterruptController

pub struct InterruptController<'a> { /* private fields */ }

An interrupt controller device node.

Interrupt controllers manage interrupt routing and handling. This type provides access to interrupt controller specific properties like the #interrupt-cells property.

Implementations

impl<'a> InterruptController<'a>

pub fn name(&self) -> &'a str

Get the name of this interrupt controller.

pub fn interrupt_cells(&self) -> Result<u8, FdtError>

Get the value of the #interrupt-cells property.

This property specifies the number of cells used to encode an interrupt specifier for this interrupt controller.