fdt_parser

Struct Header

pub struct Header {
    pub magic: u32,
    pub totalsize: u32,
    pub off_dt_struct: u32,
    pub off_dt_strings: u32,
    pub off_mem_rsvmap: u32,
    pub version: u32,
    pub last_comp_version: u32,
    pub boot_cpuid_phys: u32,
    pub size_dt_strings: u32,
    pub size_dt_struct: u32,
}

The FDT header structure.

The header is located at the start of any Device Tree Blob and contains information about the layout and version of the device tree. All multi-byte fields are stored in big-endian byte order.

Fields
magic: u32

FDT header magic number (0xd00dfeed)

totalsize: u32

Total size in bytes of the FDT structure

off_dt_struct: u32

Offset in bytes from the start of the header to the structure block

off_dt_strings: u32

Offset in bytes from the start of the header to the strings block

off_mem_rsvmap: u32

Offset in bytes from the start of the header to the memory reservation block

version: u32

FDT version

last_comp_version: u32

Last compatible FDT version

boot_cpuid_phys: u32

System boot CPU ID

size_dt_strings: u32

Length in bytes of the strings block

size_dt_struct: u32

Length in bytes of the struct block

Implementations

impl Header

pub fn from_bytes(data: &[u8]) -> Result<Self, FdtError>

Read a header from a byte slice and return an owned Header whose fields are converted from big-endian (on-disk) to host order.

Errors

Returns FdtError::BufferTooSmall if the slice is too small. Returns FdtError::InvalidMagic if the magic number is incorrect.

pub unsafe fn from_ptr(ptr: *mut u8) -> Result<Self, FdtError>

Read a header from a raw pointer and return an owned Header whose fields are converted from big-endian (on-disk) to host order.

Safety

The caller must ensure that the pointer is valid and points to a memory region of at least size_of::<Header>() bytes that contains a valid device tree blob.